requirements.txt
discord.py==2.6.0
aiohttp==3.9.4
python-dotenv==1.0.0